[update] session authorize check add

+++ main.py
@@ -17,15 +17,16 @@ def main():
         at = row['access_token']
         ats = row['access_token_secret']
         session = OAuth1Session(ck, cs, at, ats)
-        screen_name = get_user_screen_name(session)
-        print(screen_name)
+        if session.authorized:
+            screen_name = get_user_screen_name(session)
+            print(screen_name)
 
-        # result = search(screen_name, search_words, session)
-        # delete_auto_tweets(result, session)
-        tweets = get_timeline(session)
-        delete_gbf_tweets(tweets, session)
-        delete_selected_tweets(tweets, session)
-        # post_tweet("テスト", my_session)
+            # result = search(screen_name, search_words, session)
+            # delete_auto_tweets(result, session)
+            tweets = get_timeline(session)
+            delete_gbf_tweets(tweets, session)
+            delete_selected_tweets(tweets, session)
+            # post_tweet("テスト", session)
     conn.commit()
